Switch to class fixture for sql client integration tests.
